The fact that Java doesn't have unsigned numbers is actually a huge broblem. I have to use integers for the tape instead of bytes.

+++ src/main/java/info/tonyl/bf4j/lang/Environment.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Environment {
-	private static Map<Integer, Byte> cells = new HashMap<>();
+	private static Map<Integer, Integer> cells = new HashMap<>();
 	private static int pos = 0;
 	private static int boundLeft = 0;
 	private static int boundRight = 0;
@@ -37,24 +37,24 @@ public class Environment {
 		return pos;
 	}
 
-	public static byte value() {
+	public static int value() {
 		validate();
 		return cells.get(pos);
 	}
 
 	public static void increment() {
 		validate();
-		cells.put(pos, (byte) (cells.get(pos) + 1));
+		cells.put(pos, (cells.get(pos) + 1));
 	}
 
 	public static void decrement() {
 		validate();
-		cells.put(pos, (byte) (cells.get(pos) - 1));
+		cells.put(pos, (cells.get(pos) - 1));
 	}
 
 	public static void validateAt(int position) {
 		if (!cells.containsKey(pos)) {
-			cells.put(pos, (byte) 0);
+			cells.put(pos, 0);
 		}
 	}
